+++ QRayEngine/src/scenemap.h
@@ -4,10 +4,12 @@
 #define MAP_WIDTH 48
 #define MAP_HEIGHT 48
 
+#include <QtGlobal>
+
 class SceneMap
 {
 public:
-    static const int mapArray[MAP_WIDTH][MAP_HEIGHT];
+    static const quint8 mapArray[MAP_WIDTH][MAP_HEIGHT];
 };
 
 #endif // SCENEMAP_H
